# optimize Haskell implementation by adopting GHC LLVM backend and full strictness