--- a/haskell/related-post-gen.cabal
+++ b/haskell/related-post-gen.cabal
@@ -5,34 +5,30 @@ author:        thelissimus
 maintainer:    thelissimus@tuta.io
 build-type:    Simple
 
-common warnings
+common options
+  default-language: GHC2024
   ghc-options:
-    -Wall -Wextra -Weverything -Wcompat -Wno-implicit-prelude
-    -Wno-unsafe -Wno-missing-safe-haskell-mode
-    -Wno-missing-local-signatures -Wno-missing-import-lists
-    -Wno-all-missed-specialisations
+    -Wall -Wextra -fspecialise-aggressively -fllvm -optlo-O3 -O2
+    -funbox-strict-fields -fspec-constr
 
 library
-  import:           warnings
-  ghc-options:      -O2
-  hs-source-dirs:   src
-  exposed-modules:  RelatedPostGen
+  import:          options
+  hs-source-dirs:  src
+  exposed-modules: RelatedPostGen
   build-depends:
     , aeson
-    , base               >=4.7 && <5
+    , base               >=4.7     && <5
     , deepseq
-    , text
-    , unboxed-ref
+    , primitive
+    , storable-tuple
+    , text-short
     , vector
-    , vector-hashtables
-
-  default-language: GHC2021
+    , vector-hashtables  ==0.1.2.0
 
 executable related-post-gen
-  import:           warnings
-  ghc-options:      -O2
-  hs-source-dirs:   app
-  main-is:          Main.hs
+  import:         options
+  hs-source-dirs: app
+  main-is:        Main.hs
   build-depends:
     , aeson
     , base              >=4.7 && <5
@@ -42,14 +38,12 @@ executable related-post-gen
     , time
     , vector
 
-  default-language: GHC2021
-
 benchmark related-post-gen-bench
-  import:           warnings
-  ghc-options:      -with-rtsopts=-A32m -fproc-alignment=64
-  type:             exitcode-stdio-1.0
-  main-is:          Main.hs
-  hs-source-dirs:   bench
+  import:         options
+  ghc-options:    -with-rtsopts=-A32m -fproc-alignment=64
+  type:           exitcode-stdio-1.0
+  main-is:        Main.hs
+  hs-source-dirs: bench
   build-depends:
     , aeson
     , base              >=4.7 && <5
@@ -57,5 +51,3 @@ benchmark related-post-gen-bench
     , deepseq
     , related-post-gen
     , tasty-bench
-
-  default-language: GHC2021