- Switch GHC version from `9.8.1` to `9.10.1`
- Switch to LLVM backend
- Enable `Strict` extension
- `MutableByteArray` instead of `STRefU`
- `ShortText` instead of `Text` (1)
- `Storable` vector instead of `Primitive`
- Rewrite helpers as (tail-recursive) loops (2)
- Remove unnecessary warnings
- Remove bit packing
- Remove redundant standalone kind signatures

> 1. `ShortText` is actually more appropriate for the challenge since UTF-8 is enough (as opposed to `Text` with its UTF-16).
> 2. In `buildRelated` function one might see the usage of two **potentially** rule-breaking "unsafe" functions - `unsafeNew` and `unsafeFreeze`. But those names are deceptive and they are actually safe and therefore adhere to the rules. `unsafeNew` is just like `new` but doesn't initialize memory. Since we rewrite the contents anyways there is no need to initialize the vector when created. `unsafeFreeze` works in-place instead of copying the content like `freeze`. Since we won't use the accumulated mutable vector at all it is completely safe. No rules are broken.

Also, the build of `Dockerfile` might fail if there is no LLVM because Haskell implementation now depends on it. GHC supports versions from 13 up to 20 (not inclusive). [^1] Locally I used LLVM 15 and that worked fine. I wanted to test it locally but it has been building the image for hours now. The `Dockerfile` appears to include LLVM, but I haven’t verified compatibility. So, @jinyus, please double-check the Docker build.

[^1]: https://downloads.haskell.org/ghc/latest/docs/users_guide/codegens.html#llvm-code-generator-fllvm