@thelissimus 2025-02-11
Also, you might want to stick around after merging this PR because I have another version ready that utilizes `brokkr-hashtables` instead of `vector-hashtables` which has 5-10% speed gains. First let's see how `vector-hashtables` performs then `brokkr-hashtables` then we will pick the fastest one for the task. (I can't rely on my local results to make the judgement, so it is better to just run it like the rest of the languages in the benchmarks and then decide)