@jinyus 2025-05-29
Nice! This is much faster

Main:
```bash
| Haskell | 43.20 ms | 600.00 ms | 5.25 s | 5.89 s |
```

This PR:
```bash
| Haskell | 26.65 ms | 347.84 ms | 2.81 s | 3.19 s |
```